+++ src/Management/EasyAdminActionHelper.php
@@ -1,0 +1,21 @@
+<?php
+/*
+ * (c) 2026: 975L <contact@975l.com>
+ * (c) 2026: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+namespace c975L\ConfigBundle\Management;
+
+use EasyCorp\Bundle\EasyAdminBundle\Config\Action;
+use EasyCorp\Bundle\EasyAdminBundle\Config\ActionGroup;
+
+// Turns an EasyAdmin action button into icon-only, the label becoming its hover "title"
+class EasyAdminActionHelper
+{
+    public static function toIconOnly(Action|ActionGroup $action, string $title): Action|ActionGroup
+    {
+        return $action->setLabel(false)->setHtmlAttributes(['title' => $title]);
+    }
+}